refactor backend api code
